Create anchors.js
Create anchors.js file and link to it from the base template IF Site Params enableAnchors is true in config.toml. The point of this file is to override the default settings from the CDN. Use the add() method to declare which elements should have the anchor. Added article h3 - h6 on the DOMContentLoaded event.